1.3.9 for 1.21.1; - force option for particle emitting block behaviour is not available - Fire behaviour does not work - Custom TNT explosion power not configurable - Custom trident missing model generation - No per-player animation for animated models - Damage-source for custom maces might be incorrect - Biome modifications might not be able to parse biome-tags correctly

+++ src/main/java/de/tomalbrc/filament/mixin/decoration/FlowingFluidMixin.java
@@ -13,69 +13,73 @@ import net.minecraft.world.level.material.FlowingFluid;
 import net.minecraft.world.level.material.Fluid;
 import net.minecraft.world.level.material.FluidState;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 @Mixin(FlowingFluid.class)
-public abstract class FlowingFluidMixin {
-    @Shadow protected abstract void beforeDestroyingBlock(LevelAccessor level, BlockPos pos, BlockState state);
+public class FlowingFluidMixin {
 
-    @Inject(method = "canPassThroughWall(Lnet/minecraft/core/Direction;Lnet/minecraft/world/level/BlockGetter;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;)Z", at = @At("HEAD"), cancellable = true)
-    private static void filament$customCanPassThroughWall(Direction direction, BlockGetter blockGetter, BlockPos blockPos, BlockState blockState, BlockPos blockPos2, BlockState blockState2, CallbackInfoReturnable<Boolean> cir) {
-        if (DecorationRegistry.isDecoration(blockState2) || DecorationRegistry.isDecoration(blockState)) {
-            cir.setReturnValue(true);
+    @Inject(method = "canSpreadTo", at = @At("TAIL"), cancellable = true)
+    private void filament$canSpreadTo(BlockGetter blockGetter, BlockPos blockPos, BlockState blockState, Direction direction, BlockPos blockPos2, BlockState blockState2, FluidState fluidState, Fluid fluid, CallbackInfoReturnable<Boolean> cir) {
+        if (DecorationRegistry.isDecoration(blockState2)) {
+            boolean isWaterloggable = this.filament$isWaterloggable((DecorationBlock) blockState2.getBlock()) && direction != Direction.DOWN;
+            boolean isSolid = this.filament$isSolid((DecorationBlock) blockState2.getBlock()) && direction != Direction.DOWN;
+            cir.setReturnValue(isWaterloggable || !isSolid);
         }
     }
 
-    @Unique
-    private boolean filament$passes(BlockState blockState) {
-        if (DecorationRegistry.isDecoration(blockState) && (!filament$isSolid((DecorationBlock) blockState.getBlock()) || filament$isWaterloggable(blockState)))
-            return this.filament$canFlowThrough(blockState);
-        return false;
-    }
-
-    @Inject(method = "canMaybePassThrough", at = @At("RETURN"), cancellable = true)
-    private void filament$canMaybePassThrough(BlockGetter blockGetter, BlockPos blockPos, BlockState blockState, Direction direction, BlockPos blockPos2, BlockState blockState2, FluidState fluidState, CallbackInfoReturnable<Boolean> cir) {
-        // pass-thu but only non-waterloggable blocks and non-solid
-        if (filament$passes(blockState2)) cir.setReturnValue(true);
-    }
-
-    @Inject(method = "canPassThrough", at = @At("RETURN"), cancellable = true)
-    private void filament$canPassThrough(BlockGetter blockGetter, Fluid fluid, BlockPos blockPos, BlockState blockState, Direction direction, BlockPos blockPos2, BlockState blockState2, FluidState fluidState, CallbackInfoReturnable<Boolean> cir) {
-        if (filament$passes(blockState)) cir.setReturnValue(true);
-        if (filament$passes(blockState2)) cir.setReturnValue(true);
-    }
-
-    @Inject(method = "spreadTo", at = @At(value = "HEAD"))
+    @Inject(method = "spreadTo", at = @At("TAIL"))
     protected void filament$spreadTo(LevelAccessor levelAccessor, BlockPos blockPos, BlockState blockState, Direction direction, FluidState fluidState, CallbackInfo ci) {
-        if (DecorationRegistry.isDecoration(blockState) && !filament$isSolid((DecorationBlock) blockState.getBlock())) {
+        if (DecorationRegistry.isDecoration(blockState) && !filament$isWaterloggable((DecorationBlock) blockState.getBlock()) && !filament$isSolid((DecorationBlock) blockState.getBlock())) {
             if (levelAccessor.getBlockEntity(blockPos) instanceof DecorationBlockEntity decorationBlockEntity) {
                 decorationBlockEntity.destroyStructure(true);
             }
+
+            // let it overflow with liquid
+            levelAccessor.setBlock(blockPos, fluidState.createLegacyBlock(), 3);
         }
+    }
+
+    @Inject(method = "canPassThrough", at = @At("TAIL"), cancellable = true)
+    private void filament$canPassThrough(BlockGetter blockGetter, Fluid fluid, BlockPos blockPos, BlockState blockState, Direction direction, BlockPos blockPos2, BlockState blockState2, FluidState fluidState, CallbackInfoReturnable<Boolean> cir) {
+        // pass-thu but only non-waterloggable blocks and non-solid
+        if (DecorationRegistry.isDecoration(blockState2) && !filament$isWaterloggable((DecorationBlock) blockState2.getBlock()) && !filament$isSolid((DecorationBlock) blockState2.getBlock()))
+            cir.setReturnValue(this.filament$canFlowThrough(blockState2));
+    }
+    @Inject(method = "canPassThroughWall", at = @At("TAIL"), cancellable = true)
+    private void filament$canPassThroughWall(Direction direction, BlockGetter blockGetter, BlockPos blockPos, BlockState blockState, BlockPos blockPos2, BlockState blockState2, CallbackInfoReturnable<Boolean> cir) {
+        if (DecorationRegistry.isDecoration(blockState) || DecorationRegistry.isDecoration(blockState2))
+            cir.setReturnValue(true);
     }
 
     @Unique
     private boolean filament$canFlowThrough(BlockState blockState) {
-        DecorationBlock decorationBlock = (DecorationBlock) blockState.getBlock();
-        if (decorationBlock.getDecorationData() != null) {
-            return !decorationBlock.getDecorationData().hasBlocks();
+        if (DecorationRegistry.isDecoration(blockState)) {
+            DecorationBlock decorationBlock = (DecorationBlock) blockState.getBlock();
+
+            if (decorationBlock.getDecorationData() != null && !blockState.hasProperty(BlockStateProperties.WATERLOGGED)) {
+                return !decorationBlock.getDecorationData().hasBlocks() && !decorationBlock.getDecorationData().properties().solid;
+            }
         }
 
         return false;
     }
 
     @Unique
-    private static boolean filament$isWaterloggable(BlockState decorationBlock) {
-        return decorationBlock.hasProperty(BlockStateProperties.WATERLOGGED);
+    private boolean filament$isWaterloggable(DecorationBlock decorationBlock) {
+        if (decorationBlock.getDecorationData() != null) {
+            return decorationBlock.defaultBlockState().hasProperty(BlockStateProperties.WATERLOGGED);
+        }
+
+        return false;
     }
 
     @Unique
-    private static boolean filament$isSolid(DecorationBlock decorationBlock) {
+    private boolean filament$isSolid(DecorationBlock decorationBlock) {
         if (decorationBlock.getDecorationData() != null) {
             return decorationBlock.getDecorationData().properties().solid;
         }
